requirements.txt
streamlit==1.39.0
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
transformers==4.45.2
torch>=2.3.0
pymupdf==1.24.10
numpy>=1.26
scikit-learn>=1.5.0
pyyaml>=6.0.1
pdfplumber==0.7.6
nltk>=3.8.1